apply both filter and advice to invalidate cookies
