+++ src/main/java/org/folio/login/controller/ApiExceptionHandler.java
@@ -27,7 +27,7 @@ import java.util.Optional;
 import lombok.extern.log4j.Log4j2;
 import org.apache.logging.log4j.Level;
 import org.folio.cql2pgjson.exception.CQLFeatureUnsupportedException;
-import org.folio.login.controller.cookie.InvalidateCookiesOnException;
+import org.folio.login.controller.cookie.advice.InvalidateCookiesOnException;
 import org.folio.login.domain.dto.Error;
 import org.folio.login.domain.dto.ErrorCode;
 import org.folio.login.domain.dto.ErrorResponse;
