+++ src/main/java/org/folio/login/controller/cookie/filter/HttpRequestResponseHolder.java
@@ -1,4 +1,4 @@
-package org.folio.login.controller.cookie;
+package org.folio.login.controller.cookie.filter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
